All pages except server stats ready

+++ Kontur.GameStats.Server/Model/MatchResults.cs
@@ -16,7 +16,7 @@ namespace Kontur.GameStats.Server.Model
 
         public int TimeLimit { get; set; }
 
-        public double TimeElapsed { get; set; }
+        public float TimeElapsed { get; set; }
 
         public List<PlayerScore> Scoreboard { get; set; }
